Added support for configuration files
Configuration files use the same format as the --config string (except that commas can
be replaced by newlines), and can be placed in ~/.ocp/ocp-indent.conf, or in a .ocp-indent
file in any parent directory. This provides a convenient way to have per-project setups
that can be easily shared by any contributor.

Setting the ocp-indent-config variable in emacs-customise is now deprecated.