Fixed bug with 'v}' in code. (closes #61)
Oh indeed, thanks for the precise report: this was indeed a bug with the parser.
These rules are for handling verbatim blocks within ocamldoc comments, but should
of course not change the behaviour of parsing normal code (here there was a mixup
between the `]}` and `v}` cases, and the 'v' was incorrectly parsed as RBRACKET).