Improved vim indentation script
Thanks to @Daiwen, I think I finally got it right. ocp-indent should be called from the given line,
and the cache only used for consecutive lines, because ocp-indent adapts to actual indentation. Users
don't care that the line is correctly indented "with respect to how it would have indented above lines"

This is enough to get good results for 1-line and region indentation. An additional trick is needed
to detect consecutive lines because vim doesn't call the function on empty lines.

Closes #186, thanks!