No extra indent after `fun ->` when following a `let`
This provides the nice property that if you have:
```
let f = fun x ->
  foo
```
and decide to skip a line before `fun ->` (e.g. when adding type
annotations after `f`), you don't need to add extra indent to the
function body:
```
let f =
  fun x ->
  foo
```

vs. previously
```
let f =
  fun x ->
    foo
```

This is a change that can affect quite a lot of programs though. I'd
prefer to have it, but if it breaks too many programs, we can add an
option.

Closes #188

Note: expect tests needed to be adjusted in a few places