Fix some warnings
Warning 57 ("ambiguous or-pattern under guard") is made no-error; we
can't disable it locally because we want to stay backwards-compatible
with very old versions of OCaml, and we don't want to disable the
warning because it has caught meaningful things in the past.

Also it's not very easy to work-around without duplicating things.